SlackBot/SupremeCourtAI/requirements.txt
plotly
scipy
python-dotenv
flask_executor
slack_sdk
openai
numpy
pandas